feat: use const for default erl activation period value
The reason for prepending the variable/argument with
`erl_` is to avoid confusion in `utils.js` file where
a generic `normalizeTemporals` function is defined.

We don't want other developers to think that this applies
to anything other than erl.

Another way to make this explicit would be to call
`normalizeTemporals` with another argument that indicates
we're resolving erl config at the moment, so that it
considers activation_period_mins only then.